refactor: refresh button
