+++ apps/dashboard/app/(app)/ratelimits/[namespaceId]/logs/components/controls/components/logs-refresh.tsx
@@ -1,64 +1,25 @@
-import { useKeyboardShortcut } from "@/hooks/use-keyboard-shortcut";
+import { RefreshButton } from "@/components/logs/refresh-button";
 import { trpc } from "@/lib/trpc/client";
-import { Refresh3 } from "@unkey/icons";
-import { Button } from "@unkey/ui";
-import { cn } from "@unkey/ui/src/lib/utils";
-import { useState } from "react";
 import { useRatelimitLogsContext } from "../../../context/logs";
 import { useFilters } from "../../../hooks/use-filters";
 
-const REFRESH_TIMEOUT_MS = 1000;
-
-export const LogsRefresh = () => {
-  const { isLive, toggleLive } = useRatelimitLogsContext();
+export const RatelimitLogsRefresh = () => {
+  const { toggleLive, isLive } = useRatelimitLogsContext();
   const { filters } = useFilters();
   const { ratelimit } = trpc.useUtils();
-  const [isLoading, setIsLoading] = useState(false);
-  const [refreshTimeout, setRefreshTimeout] = useState<NodeJS.Timeout | null>(null);
-
   const hasRelativeFilter = filters.find((f) => f.field === "since");
-  useKeyboardShortcut("r", () => {
-    hasRelativeFilter && handleSwitch();
-  });
 
-  const handleSwitch = () => {
-    if (isLoading) {
-      return;
-    }
-
-    const isLiveBefore = Boolean(isLive);
-    setIsLoading(true);
-    toggleLive(false);
+  const handleRefresh = () => {
     ratelimit.logs.query.invalidate();
     ratelimit.logs.queryRatelimitTimeseries.invalidate();
-
-    if (refreshTimeout) {
-      clearTimeout(refreshTimeout);
-    }
-    const timeout = setTimeout(() => {
-      setIsLoading(false);
-      if (isLiveBefore) {
-        toggleLive(true);
-      }
-    }, REFRESH_TIMEOUT_MS);
-    setRefreshTimeout(timeout);
   };
 
   return (
-    <Button
-      onClick={handleSwitch}
-      variant="ghost"
-      className={cn(
-        "px-2 text-accent-12 [&_svg]:text-accent-9",
-        "relative overflow-hidden",
-        "transition-transform",
-        isLoading && "bg-gray-3",
-      )}
-      disabled={!hasRelativeFilter || isLoading}
-    >
-      {isLoading && <div className="absolute inset-0 bg-accent-6 animate-fill-left" />}
-      <Refresh3 className="size-4 relative z-10" />
-      <span className="font-medium text-[13px] relative z-10">Refresh</span>
-    </Button>
+    <RefreshButton
+      onRefresh={handleRefresh}
+      isEnabled={Boolean(hasRelativeFilter)}
+      isLive={isLive}
+      toggleLive={toggleLive}
+    />
   );
 };
